+++ src/Interfaces/Values/AlfEmptyWork.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Alf\Interfaces\Values;
+
+interface AlfEmptyWork extends AlfEmptyGet, AlfEmptySet {
+
+}