+++ src/Interfaces/Values/AlfNullOrEmptyWorkTrait.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Alf\Interfaces\Values;
+
+use Alf\Attributes\AlfAttrAutoComplete;
+use Alf\Attributes\AlfAttrTraitAutoCall;
+use Alf\Services\AlfProgramming;
+use JetBrains\PhpStorm\Pure;
+
+trait AlfNullOrEmptyWorkTrait {
+
+    use AlfNullWorkTrait, AlfEmptyWorkTrait;
+
+    /** @AlfAttrAutoComplete */
+    #[AlfAttrAutoComplete]
+    public static function _AlfNullOrEmptyWork($obj) : AlfNullOrEmptyWork {
+        return AlfProgramming::_()->unused($obj, static::_AlfNullWork($obj), static::_AlfEmptyWork($obj));
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfNullOrEmptyWorkTraitCTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfNullOrEmptyWorkTraitDTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfNullOrEmptyWorkTraitClone() : void {
+    }
+
+    #[Pure]
+    public function isNullOrEmpty() : bool {
+        return $this->isNull() || $this->isEmpty();
+    }
+
+}