+++ tests/Basic/InterfaceTest.php
@@ -1,7 +1,7 @@
 <?php
 
 test('interface has trait',
-    function (string $interfaceName) {
+    function (string $interfaceName) : void {
 
         $needTraitName = $interfaceName.'Trait';
         expect(trait_exists($needTraitName))->toBeTrue();