+++ tests/Basic/ParameterTest.php
@@ -1,7 +1,9 @@
 <?php
 
 test('parameter-dummy',
-    function (string $className) {
+    function (string $className) : void {
+
         expect(true)->toBeTrue();
+
     })->with(listAlfAll());