+++ tests/Pest.php
@@ -3,6 +3,9 @@
 // uses(Tests\TestCase::class)->in('Feature');
 // expect()->extend('toBeOne', function () { return $this->toBe(1); });
 
+// ToDo: Attribute AlfAttrParamsIsInt
+// ToDo: autocomplete for interface in traits
+
 use Alf\AlfBasicAttribute;
 use Alf\AlfBasicClass;
 use Alf\AlfBasicSingleton;
@@ -16,8 +19,12 @@ use Alf\Interfaces\Values\AlfEmptyGet;
 use Alf\Interfaces\Values\AlfEmptyGetTrait;
 use Alf\Interfaces\Values\AlfEmptySet;
 use Alf\Interfaces\Values\AlfEmptySetTrait;
+use Alf\Interfaces\Values\AlfEmptyWork;
+use Alf\Interfaces\Values\AlfEmptyWorkTrait;
 use Alf\Interfaces\Values\AlfNullGet;
 use Alf\Interfaces\Values\AlfNullGetTrait;
+use Alf\Interfaces\Values\AlfNullOrEmptyWork;
+use Alf\Interfaces\Values\AlfNullOrEmptyWorkTrait;
 use Alf\Interfaces\Values\AlfNullSet;
 use Alf\Interfaces\Values\AlfNullSetTrait;
 use Alf\Interfaces\Values\AlfNullWork;
@@ -39,7 +46,9 @@ function listAlfInterfaces() : array {
         // Interfaces/Values
         AlfEmptyGet::class,
         AlfEmptySet::class,
+        AlfEmptyWork::class,
         AlfNullGet::class,
+        AlfNullOrEmptyWork::class,
         AlfNullSet::class,
         AlfNullWork::class,
         AlfValueGet::class,
@@ -52,7 +61,9 @@ function listAlfTraits() : array {
         // Interfaces/Values
         AlfEmptyGetTrait::class,
         AlfEmptySetTrait::class,
+        AlfEmptyWorkTrait::class,
         AlfNullGetTrait::class,
+        AlfNullOrEmptyWorkTrait::class,
         AlfNullSetTrait::class,
         AlfNullWorkTrait::class,
         AlfValueGetTrait::class,
@@ -98,4 +109,58 @@ function listAlfClassesAndSingletons() : array {
 #[Pure]
 function listAlfAll() : array {
     return array_merge(listAlfClasses(), listAlfSingletons(), listAlfInterfaces(), listAlfTraits());
+}
+
+#[Pure]
+function getIntValues() : array {
+    return [
+        [
+            'set'      => null,
+            'isNull'   => true,
+            'isEmpty'  => false,
+            'get'      => 0,
+            'getValue' => null,
+        ],
+        [
+            'set'      => 0,
+            'isNull'   => false,
+            'isEmpty'  => true,
+            'get'      => 0,
+            'getValue' => 0,
+        ],
+        [
+            'set'      => 5,
+            'isNull'   => false,
+            'isEmpty'  => false,
+            'get'      => 5,
+            'getValue' => 5,
+        ],
+        [
+            'set'      => -5,
+            'isNull'   => false,
+            'isEmpty'  => false,
+            'get'      => -5,
+            'getValue' => -5,
+            'AlfInt8U' => [
+                'isEmpty'  => true,
+                'get'      => 0,
+                'getValue' => 0,
+            ],
+        ],
+        [
+            'set'      => 500,
+            'isNull'   => false,
+            'isEmpty'  => false,
+            'get'      => 500,
+            'getValue' => 500,
+            'AlfInt8U' => [
+                'get'      => 255,
+                'getValue' => 255,
+            ],
+            'AlfInt8'  => [
+                'get'      => 128,
+                'getValue' => 128,
+            ],
+        ],
+    ];
 }